+++ CHANGELOG.md
@@ -11,8 +11,10 @@ $ npm install @ngxs/store@dev
 * Feature: Add possible inheritance of state options [#750](https://github.com/ngxs/store/pull/750)
 * Feature: Add new lifecycle hook `ngxsAfterBootstrap` [#753](https://github.com/ngxs/store/pull/753)
 * Feature: Add ability to change the Execution strategy (decouple from zone.js) [#811](https://github.com/ngxs/store/pull/811)
+* Feature: Add state operators: `patch, updateItem, removeItem, insertItem, append, compose` [#799](https://github.com/ngxs/store/pull/799/)
 * Fix: Expose `ActionCompletion` [#752](https://github.com/ngxs/store/pull/752)
 * Fix: Throw error when found duplicate state names [#791](https://github.com/ngxs/store/pull/791)
+* Fix: Bind context to the selector function [#818](https://github.com/ngxs/store/pull/818)
 
 ## NGXS-Labs